Move away from Google+ API usage in google-oauth2 backend
Google+ APIs will be shut down on March 7, 2019. More details here
https://developers.google.com/+/api-shutdown

> This will be a progressive shutdown, with intermittent failures
> starting as early as January 28, 2019.

Replacing the G+ endpoint with
https://www.googleapis.com/oauth2/v3/userinfo works for the
google-oauth2 backend with almost no extra changes.

This new endpoint is currently used by
[Google's OpenID Connect](https://developers.google.com/identity/protocols/OpenIDConnect)
implementation as OpenID Connect `userinfo_endpoint`. It might not be
the best idea to hard code it like that in the long term. But this was
the one endpoint that is actually mentioned in some Google
documentation as far as I could find. Please feel free to propose better
alternatives if you can find them.

`GooglePlusAuth` backend is affected by this change too.
However according to https://developers.google.com/+/web/signin/
the Google+ Sign-in feature is also fully deprecated and will be shut
down on March 7, 2019. I'd expect for the changes related
dropping/deprecating `GooglePlusAuth` backend in this library to be done
in a separate change. This change makes sure `google-oauth2` backend
will continue to work during and after Google+ APIs deprecation.

After this change this note from the docs can be removed too I believe:

> the application requires Google+ API to be enabled in the Google
> console dashboard

At least as far as the `google-oauth2` backend goes.