Fix old fields from FIELDS_STORED_IN_SESSION persisting in session
Closes #425.
If a field from FIELDS_STORED_IN_SESSION was present in the first
request to login, it would get saved in the session and on the next
login request, if it was made without that field, it would still be
in the session despite being undesirable.
At the beginning of auth we should ensure those fields that aren't
in the request won't be present in the session.