Bump Facebook Graph API version to version 8.0
This is currently the latest version of the API available.

Each API version is available for 2 years, and the previous version 3.2 API is
reaching end of life on May 4th, 2021. It would be good to upgrade to a newer
version before then in order to have time to find any version incompatibilities
before the current version is deprecated, although I haven't been able to find
any breaking changes by looking at the version history between the versions:
https://developers.facebook.com/docs/graph-api/changelog