Fix kid lookups for key rotation (OpenID Connect)
According to the OpenID Connect Documentation
https://openid.net/specs/openid-connect-core-1_0.html#RotateSigKeys keys
should be re-fetched if a kid in an ID Token is not part of the cached
keys.

The only thing that needs to be done is to check if one of the keys
matches the kid and if not we should refetch the keys.

Fixes #591