Add Facebook Limited Login backend (#590)
* Add backend for Facebook Limited Login

This is a new login method for Facebook which uses OpenID Connect and doesn't
have the ability of associating a user logging into one app with the IDFA token
on iOS devices:
https://developers.facebook.com/docs/facebook-login/limited-login.

* Add entry to changelog

* Use the decoded JWT as the "response" data going through the pipeline

This is for pipelines to work correctly so they don't need to decode the JWT at
each step in order to read the user information.

* There's no access token to call APIs to get user data, so make it a no-op

* Remove validation of nonce as it's not generated server-side

The client (iOS app or Unity for now) will have to generate the nonce and check
it matches the response from Facebook. On the Python server-side we can't check
the nonce as it wasn't generated by us so skip the check.

* JWT is valid for an hour

* Handle longer expiration times better in test_invalid_issue_time

Using a hardcoded value of 1 hour doesn't work well for backends that has a 1
hour (or longer) expiration period. Using double the expiration period
configured on the backend should ensure the test work regardless of experiation
period set on the backend.

* Correct import path to OpenIdConnectTestMixin