Pass "next" URL through SAML RelayState
This is the primary use case for RelayState, but the SAML backend did not
previously support this use. Instead, it relied on the Strategy to store
the "next" URL in session state and restore it from there after login.
Unfortunately, in the case of SAML, this does not work if the server has
set the session cookie's `SameSite` attribute to "Lax" or "Strict", as the
SAML POST request from the IdP will not contain the session cookie.

The new `RelayState` format (a JSON object) allows for further
extensibility, as arbitrary additional fields can be added in the future.