fix: update base URL to allow custom domains
fixes #809.

The Authority Host can be defined to be a custom domain as well using front door in Azure. With this, the tenant name is not prefixed to the authority host in the base url.

We update this by prefixing the tenant name with the default authority host only if the authority host is not defined.