feat: allow to configure JWT leeway in open_id_connect and increase default to 1 second (#1031)
* feat: allow to configure JWT leeway in open_id_connect

The default value defined in the jwt module is 0.

* feat: increase default JWT_LEEWAY in open_id_connect to 1.0 seconds

The default value defined in the jwt module is 0 seconds, so when the
time is off just by a few milliseconds, the token validation fails with
error "The token is not yet valid (iat)".

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------

Co-authored-by: Michal Čihař <michal@weblate.org>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>