feat: send user agent by default
This is now needed by mediawiki.py (Wikimedia policy has required it for
some 15 years [1], and more recently has enforced it more strictly, see
T400119 [2]), but discussion in #1241 concluded that it should really be
enabled by default for all sites. If it turns out to break anything,
that particular backend can override it to False again.

[1]: https://lists.wikimedia.org/hyperkitty/list/wikitech-l@lists.wikimedia.org/message/R4RU7XTBM5J3BTS6GGQW77NYS2E4WGLI/
[2]: https://phabricator.wikimedia.org/T400119