fix(backens): avoid passing pipeline data as kwargs
The pipeline can be customized to contain any arbitrary data and that
could result in passing duplicate parameters to the extra_data method.
Pasing the pipeline kwargs in a separate dict avoids that. Passing
positional args is dropped as that can break in various ways.

This is possibly a breaking change for third-party backends.